requirements.txt
asgiref==3.8.1
Django==3.2.23
djangorestframework==3.12.0
pytz==2024.2
sqlparse==0.5.2
typing_extensions==4.12.2
